+++ site/plugins/komments/lib/WebmentionReceiver.php
@@ -38,6 +38,7 @@ class WebmentionReceiver
             authorAvatar: $webmention['author']['avatar'] ?? '',
             authorEmail: null,
             authorUrl: $webmention['source'],
+            verification_status: $autoPublish ? 'PUBLISHED' : 'PENDING',
             published: $autoPublish,
             verified: false,
             spamlevel: $spamlevel,