+++ site/plugins/komments/migrations/database_202511122120.sql
@@ -1,0 +1,2 @@
+ALTER TABLE comments DROP COLUMN status;
+ALTER TABLE comments ADD COLUMN verification_status TEXT DEFAULT "PENDING";